requirements.txt
# -*- coding: utf-8 -*-
"""
Created on Sun Oct 19 15:00:12 2025

@author: mthom
"""

flask
flask-cors
python-dotenv
sqlite3-builtin
